chore: move shared theme to shared ui
